+++ examples/data-sources/nxos_hmm_instance/data-source.tf
@@ -1,2 +1,0 @@
-data "nxos_hmm_instance" "example" {
-}