+++ examples/data-sources/nxos_hmm_interface/data-source.tf
@@ -1,3 +1,0 @@
-data "nxos_hmm_interface" "example" {
-  interface_id = "vlan10"
-}