+++ examples/resources/nxos_hmm_instance/import-by-identity.tf
@@ -1,5 +1,0 @@
-import {
-  to = nxos_hmm_instance.example
-  identity = {
-  }
-}