+++ examples/resources/nxos_hmm_instance/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_hmm_instance.example
-  id = ""
-}