+++ examples/resources/nxos_hmm_instance/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_hmm_instance.example ""