+++ examples/resources/nxos_hmm_instance/resource.tf
@@ -1,4 +1,0 @@
-resource "nxos_hmm_instance" "example" {
-  admin_state = "enabled"
-  anycast_mac = "20:20:00:00:10:10"
-}