+++ examples/resources/nxos_hmm_interface/import-by-identity.tf
@@ -1,6 +1,0 @@
-import {
-  to = nxos_hmm_interface.example
-  identity = {
-    "interface_id" : "<interface_id>"
-  }
-}