+++ examples/resources/nxos_hmm_interface/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_hmm_interface.example
-  id = "<interface_id>"
-}