+++ examples/resources/nxos_hmm_interface/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_hmm_interface.example "<interface_id>"