+++ examples/resources/nxos_hmm_interface/resource.tf
@@ -1,5 +1,0 @@
-resource "nxos_hmm_interface" "example" {
-  interface_id = "vlan10"
-  admin_state  = "enabled"
-  mode         = "anycastGW"
-}