+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - BREAKING CHANGE: Omit empty or null attribute values from API payload by default
 - BREAKING CHANGE: Remove `nxos_default_qos_class_map_dscp` resource and data source, DSCP values are now managed as child classes of `nxos_default_qos_class_map`
+- BREAKING CHANGE: Remove `nxos_hmm_instance` resource and data source, instance configuration is now managed as a child of `nxos_hmm`
+- BREAKING CHANGE: Remove `nxos_hmm_interface` resource and data source, interfaces are now managed as child classes of `nxos_hmm`
 - Add `nxos_features` resource and data source
 
 ## 0.7.0